Fix ai-integration compilation and form_detector test
Fixed:
- search.rs: Error type conversion to SlapperError (Parse, Network variants)
- session.rs: Fixed cookies extraction using manual Set-Cookie header parsing
- session.rs: Fixed TargetType import and csrf token lookup
- session.rs: Fixed StatusCode parse type annotation
- session.rs: Fixed form_re duplicate ok()? bug
- session.rs: Fixed has_field_named to use find_field_name (regex-based)
- session.rs: Fixed NotFound error variant usage
- fuzzer.rs: Added missing FuzzArgs fields (calibrate, fc, fs, fw, fl, ft, fr)

Clippy auto-fix:
- Added Default impl for 8 types (8 structs from plan B.3)

Test results:
- 1115 base tests pass
- 1238 ai-integration tests pass (7 pre-existing AI test failures remain)

+++ crates/slapper/src/proxy/health.rs
@@ -93,7 +93,7 @@ impl HealthChecker {
             reqwest::Proxy::all(&proxy_url)?
         };
 
-        let timeout_secs = (self.config.timeout_ms as u64) / 1000;
+        let timeout_secs = self.config.timeout_ms / 1000;
         let client = create_insecure_client_with_options(timeout_secs, |builder| {
             builder.proxy(reqwest_proxy)
         })?;
